review: rename files and directory
