+++ apps/docs/src/public/apple-touch-icon.png
@@ -1,1 +1,1 @@
-../../../../packages/assets/src/logos/onyx-192x192.png
+../../../../packages/assets/src/onyx-brand/signet-192x192.png
