+++ apps/docs/src/public/images/og-image.png
@@ -1,0 +1,1 @@
+../../../../../packages/assets/src/onyx-brand/og-image.png
