--- apps/docs/src/public/images/og-logo.png
@@ -1,1 +1,0 @@
-../../../../../packages/assets/src/logos/og-logo.png
