+++ apps/docs/src/public/web-app-manifest-192x192.png
@@ -1,1 +1,1 @@
-../../../../packages/assets/src/logos/onyx-192x192.png
+../../../../packages/assets/src/onyx-brand/signet-192x192.png
