+++ apps/docs/src/public/web-app-manifest-512x512.png
@@ -1,1 +1,1 @@
-../../../../packages/assets/src/logos/onyx-512x512.png
+../../../../packages/assets/src/onyx-brand/signet-512x512.png
