+++ packages/assets/src/onyx-brand/og-image.png
