+++ packages/assets/src/onyx-brand/signet-192x192.png
