+++ packages/assets/src/onyx-brand/signet-512x512.png
