+++ packages/sit-onyx/.storybook/preview.ts
@@ -11,8 +11,8 @@ import { createOnyx } from "../src";
 import { a11yTags } from "../src/a11yConfig";
 import docsTemplate from "./docs-template.mdx";
 import { enhanceFormInjectedSymbol } from "./formInjected";
-import logoDark from "@sit-onyx/assets/logos/onyx-dark.svg";
-import logoLight from "@sit-onyx/assets/logos/onyx-light.svg";
+import logoDark from "@sit-onyx/assets/onyx-brand/logo-on-dark.svg";
+import logoLight from "@sit-onyx/assets/onyx-brand/logo-on-light.svg";
 import { enhanceSkeletonInjectedSymbol } from "./skeletonInjected";
 
 const enabledRules = getRules(a11yTags).map((ruleMetadata) => ({
