+++ packages/sit-onyx/README.md
@@ -1,5 +1,5 @@
 <div align="center" style="text-align: center">
-<img alt="Onyx logo" src="https://raw.githubusercontent.com/SchwarzIT/onyx/main/packages/assets/src/logos/favicon.svg" height="96px">
+<img alt="Onyx logo" src="https://raw.githubusercontent.com/SchwarzIT/onyx/main/packages/assets/src/onyx-brand/signet.svg" height="96px">
 </div>
 
 <br>
