+++ packages/sit-onyx/src/components/OnyxNavBar/examples/VerticalExample.vue
@@ -23,7 +23,7 @@ import {
   OnyxUserMenu,
   useResizeObserver,
 } from "../../../index.js";
-import logoUrl from "@sit-onyx/assets/logos/signet.svg";
+import logoUrl from "@sit-onyx/assets/onyx-brand/signet.svg";
 
 const expanded = ref(false);
