+++ packages/sit-onyx/src/components/examples/ComponentShowcase/ComponentShowcase.vue
@@ -7,7 +7,7 @@ import {
   iconShareIos,
 } from "@sit-onyx/icons";
 import { ref } from "vue";
-import logoUrl from "@sit-onyx/assets/logos/signet.svg";
+import logoUrl from "@sit-onyx/assets/onyx-brand/signet.svg";
 import OnyxAvatar from "../../OnyxAvatar/OnyxAvatar.vue";
 import OnyxAvatarStack from "../../OnyxAvatarStack/OnyxAvatarStack.vue";
 import OnyxBadge from "../../OnyxBadge/OnyxBadge.vue";
